requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pywebview>=5.1
mlx-whisper>=0.4.0
mlx-lm>=0.16.0
silero-vad>=5.1.2
sounddevice>=0.4.7
numpy>=1.26.0
py2app>=0.28.7
